+++ packages/components/organisms/f-footer/CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.12.0
+------------------------------
+*March 18, 2021*
+
+## Added
+- Courier links and country selector component tests for tenants: 'ie', 'nz', 'dk', 'es', 'it', 'no', 
+- Canada(FR) to dropdown country selector list
+- Accessibility tests for the above tenants
+
+
 v4.11.0
 ------------------------------
 *March 15, 2021*